requirements.txt
# ARA FULL BOT V2 - Dependencies

# Data Fetching
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Charting
mplfinance>=0.12.9b7
matplotlib>=3.7.0

# Telegram
requests>=2.31.0

# Database
# SQLite is built-in, PostgreSQL optional:
# psycopg2-binary>=2.9.0

# ML (Optional)
# scikit-learn>=1.3.0
# lightgbm>=4.0.0

# Utilities
python-dotenv>=1.0.0  # Required for .env file support

